Created separate extensions for LottieView
