+++ Package.swift
@@ -6,7 +6,8 @@ import PackageDescription
 let package = Package(
     name: "LottieUI",
     platforms: [
-        .iOS(.v13)
+        .iOS(.v13),
+        .macOS(.v10_15)
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
