+++ Sources/LottieUI/AsyncLottieView/AsyncLottieView.swift
@@ -22,21 +22,22 @@ public struct AsyncLottieView<Content: View>: View {
     ///   - animationCache: Cache to improve performance when playing recurrent animations.
     ///   - animation: A closure that takes a `LottieView` as an input and returns a view with the animation to be displayed. You can modify the animation view as needed before running it.
     ///   - placeholder: A closure that returns the view to be displayed until the Lottie is successfully downloaded.
-    public init<L, P>(
-        url: URL,
-        animationCache: AnimationCacheProvider? = LRUAnimationCache.sharedCache,
-        @ViewBuilder animation: @escaping (LottieView) -> L,
-        @ViewBuilder placeholder: @escaping () -> P
-    ) where Content == _ConditionalContent<L, P>, L: View, P: View {
-        self.url = url
-        self.cacheProvider = animationCache
-        self.content = { phase -> _ConditionalContent<L, P> in
-            switch phase {
-            case .success(let lottieView): return ViewBuilder.buildEither(first: animation(lottieView))
-            default: return ViewBuilder.buildEither(second: placeholder())
-            }
-        }
-    }
+    // TODO: - Remove private API code (_ConditionalContent)
+//    public init<L, P>(
+//        url: URL,
+//        animationCache: AnimationCacheProvider? = LRUAnimationCache.sharedCache,
+//        @ViewBuilder animation: @escaping (LottieView) -> L,
+//        @ViewBuilder placeholder: @escaping () -> P
+//    ) where Content == _ConditionalContent<L, P>, L: View, P: View {
+//        self.url = url
+//        self.cacheProvider = animationCache
+//        self.content = { phase -> _ConditionalContent<L, P> in
+//            switch phase {
+//            case .success(let lottieView): return ViewBuilder.buildEither(first: animation(lottieView))
+//            default: return ViewBuilder.buildEither(second: placeholder())
+//            }
+//        }
+//    }
     
     /// Creates a view that presents a Lottie animation from a remote URL to be displayed in phases
     /// - Parameters:
